+++ werkit/compute/graph/__init__.py
@@ -1,3 +1,4 @@
-from ._types import Input, assert_valid_type
-from ._decorators import intermediate, output, Bound
-from ._state_manager import DependencyGraph, StateManager
+from ._types import assert_valid_type
+from ._binding import bind_to_state_manager
+from ._state_manager import StateManager
+from ._dependency_graph import DependencyGraph, Input, InnerNode, Intermediate, Output, intermediate, output
